# fix reindex optimizeStorage=ALL_VERSIONS