--- a/hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/dao/BaseHapiFhirSystemDao.java
+++ b/hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/dao/BaseHapiFhirSystemDao.java
@@ -205,7 +205,7 @@ public abstract class BaseHapiFhirSystemDao<T extends IBaseBundle, MT> extends B
 			 *
 			 * However, for realistic average workloads, this should reduce the number of round trips.
 			 */
-			if (idChunk.size() >= 2) {
+			if (!idChunk.isEmpty()) {
 				List<ResourceTable> entityChunk = prefetchResourceTableHistoryAndProvenance(idChunk);
 
 				if (thePreFetchIndexes) {